Install colorio Python package

This function wraps installation functions from reticulate to install the Python package colorio.

install_colorio(
  method = c("conda", "virtualenv"),
  envname = "r-reticulate",
  ...
)

Arguments

method
character, indicates to use "conda" or "virtualenv"

envname
character, name of environment into which to install

...
other arguments sent to reticulate::conda_install() or reticulate::virtualenv_install()

Value

Invisible NULL, called for side effects

Details

This package uses the reticulate package to make an interface with the colorio Python package. To promote consistency in usage of reticulate among different R packages, it is recommended to use a common Python environment, called "r-reticulate".

Depending on your setup, you can create this environment using reticulate::conda_create() or reticulate::virtualenv_create(), as described in this reticulate article.

Examples

if (FALSE) {
  # not run because it requires Python
  install_colorio()
}